requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Security
Werkzeug==3.0.1
itsdangerous==2.1.2

# Machine Learning (using flexible versions for compatibility)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Email Analysis
dkimpy>=1.1.0
dnspython>=2.4.0
pyspf>=2.0.14

# File Analysis (Windows-compatible version)
python-magic-bin==0.4.14
pefile>=2023.2.7

# Reporting
reportlab>=4.0.0
Pillow>=10.0.0

# External APIs
requests>=2.31.0

# Visualization
matplotlib>=3.8.0

# Database
Flask-Migrate==4.0.5

# Utilities
python-dotenv>=1.0.0
